money의 amount 데이터타입 변경. money equals test. 환전해야함.

+++ src/main/java/com/nhnacademy/project/exception/InvalidMoneyAmountException.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.project.exception;
 
 public class InvalidMoneyAmountException extends RuntimeException {
-    public InvalidMoneyAmountException(int amount) {
+    public InvalidMoneyAmountException(double amount) {
         super("Invalid money amount: " + amount);
     }
 }
